[clang-tidy] Made mesos-this-capture check tighter.
Improve the `mesos-this-capture` to remove fall positives. This was
triggered by the following example,

    https://github.com/apache/mesos/blob/d2117362349ab4c383045720f77d42b2d9fd6871/src/slave/containerizer/mesos/io/switchboard.cpp#L1487

In its old form we did not make sure that the found undeferred lambda
actually captured `this`, but instead triggered on any lambda containing
some `CXXThisExpr` in any descendants. This caused problems e.g., in the
case of nested lambdas where one was undeferred while the other one
capturing a `this`. With this change we make sure to check the correct
lambda for `this`-capture.

The matcher also did not make sure that the matched `CxxMemberCallExpr`s
where actually on some process `Future`. Added an `on` narrower to
ensure this.