[clang-tidy] Added checker catching some ignored return types.
This checker is intended to catch instances where a function returning
e.g., a `Try<Nothing>` is called, but the return value is not examined
for errors.

The checker is added under the `alpha-mesos` prefix as it produces *a
lot* of diagnostics and e.g., not enabled in the mesos-tidy setup. This
is to make sure we do not hide other warning types at the moment.