[clang-tidy] Tightened mesos-this-capture check.
To remove false positives, this patch replaces uses of unconstrained
transversal matchers like `hasDescendent` or `hasAncestor` with direct
matches on the actual ASTs. This makes whitelisting "safe" dispatchers
unnecessary as they lead to other unambiguously different trees.

With this we can introduce matchers for lambdas which are only
referenced via a variable.