Only check this captures in pid-less version of `process::loop`.
We were previously analyzing both the overload of process::loop taking a
PID and the one without it. While the PID-less version dispatches
internally to an ad hoc process, the version taking a PID already
ensured itself that it dispatches to process of the passed PID. This
lead to a number of false positives.

This patch tigthens the matcher so only the PID-less overload is
analyzed which removes a huge number of false positives.